requirements.txt
Flask==2.3.3
google-generativeai==0.3.2
markdown==3.5.1